+++ pandas_datareader/datareaders/google_finance_quotes.py
@@ -1,2 +1,10 @@
 def get_quote_google(symbols):
-    raise NotImplementedError("Google Finance doesn't have this functionality")
+    """
+    Get current yahoo quote
+
+    (Should) Returns a DataFrame
+
+    ToDo: Not implemented
+    """
+    msg = "Google Finance doesn't have this functionality - can't get quote for %s" % symbols
+    raise NotImplementedError(msg)
