+++ pandas_datareader/datareaders/yahoo_finance.py
@@ -41,8 +41,8 @@ def get_data_yahoo(symbols=None, start=None, end=None, retry_count=3,
     """
     if interval not in ['d', 'w', 'm', 'v']:
         raise ValueError("Invalid interval: valid values are 'd', 'w', 'm' and 'v'")
-    hist_data = _get_data_from(symbols, start, end, interval, retry_count, pause,
-                          chunksize, 'yahoo')
+    hist_data = _get_data_from(symbols, start, end, interval, retry_count, pause, \
+                    chunksize, 'yahoo')
     if ret_index:
         hist_data['Ret_Index'] = _calc_return_index(hist_data['Adj Close'])
     if adjust_price:
