+++ pandas_datareader/datareaders/yahoo_finance_daily.py
@@ -1,5 +1,5 @@
-from pandas.io.common import urlencode
-from pandas_datareader.data import _sanitize_dates, _retry_read_url
+from pandas_datareader.url_request import _retry_read_url
+from pandas_datareader.date_chunks import _sanitize_dates
 
 _HISTORICAL_YAHOO_URL = 'http://ichart.finance.yahoo.com/table.csv?'
