+++ pandas_datareader/datareaders/yahoo_finance_options.py
@@ -1,7 +1,12 @@
+import warnings
 import datetime as dt
+import numpy as np
 from pandas.io.html import read_html
 from pandas import to_datetime
 from pandas import concat, DatetimeIndex, Series
+from pandas.tseries.offsets import MonthEnd
+from pandas.util.testing import _network_error_classes
+from pandas_datareader.shared import RemoteDataError
 
 # Items needed for options class
 CUR_MONTH = dt.datetime.now().month
