+++ pandas_datareader/shared.py
@@ -1,10 +1,22 @@
+import warnings
+import numpy as np
 import pandas.compat as compat
+from pandas.core.common import PandasError
+from pandas import Panel, DataFrame
+
+from .date_chunks import _in_chunks
 
 from .datareaders.google_finance_daily import _get_hist_google
 from .datareaders.yahoo_finance_daily import _get_hist_yahoo
 
-_source_functions = {'google': _get_hist_google, 'yahoo': _get_hist_yahoo}
 
+class SymbolWarning(UserWarning):
+    pass
+
+class RemoteDataError(PandasError, IOError):
+    pass
+
+_source_functions = {'google': _get_hist_google, 'yahoo': _get_hist_yahoo}
 
 def _get_data_from(symbols, start, end, interval, retry_count, pause,
                    chunksize, source):
@@ -22,3 +34,34 @@ def _get_data_from(symbols, start, end, interval, retry_count, pause,
         hist_data = _dl_mult_symbols(symbols, start, end, interval, chunksize,
                                      retry_count, pause, src_fn)
     return hist_data
+
+def _dl_mult_symbols(symbols, start, end, interval, chunksize, retry_count, pause,
+                     method):
+    stocks = {}
+    failed = []
+    passed = []
+    for sym_group in _in_chunks(symbols, chunksize):
+        for sym in sym_group:
+            try:
+                stocks[sym] = method(sym, start, end, interval, retry_count, pause)
+                passed.append(sym)
+            except IOError:
+                warnings.warn('Failed to read symbol: {0!r}, replacing with '
+                              'NaN.'.format(sym), SymbolWarning)
+                failed.append(sym)
+
+    if len(passed) == 0:
+        raise RemoteDataError("No data fetched using "
+                              "{0!r}".format(method.__name__))
+    try:
+        if len(stocks) > 0 and len(failed) > 0 and len(passed) > 0:
+            df_na = stocks[passed[0]].copy()
+            df_na[:] = np.nan
+            for sym in failed:
+                stocks[sym] = df_na
+        return Panel(stocks).swapaxes('items', 'minor')
+    except AttributeError:
+        # cannot construct a panel with just 1D nans indicating no data
+        raise RemoteDataError("No data fetched using "
+                              "{0!r}".format(method.__name__))
+
