+++ pandas_datareader/tests/test_data.py
@@ -24,8 +24,8 @@ except ImportError:
     from urllib2 import HTTPError
 
 import pandas_datareader.data as web
-from pandas_datareader.data import (
-    DataReader, SymbolWarning, RemoteDataError)
+from pandas_datareader.data import DataReader
+from pandas_datareader.shared import SymbolWarning, RemoteDataError
 from pandas_datareader.datareaders.yahoo_finance_quotes import _yahoo_codes
 
 def _skip_if_no_lxml():
